// server/lib/zitadel/proto/zitadel/settings/v2/legal_settings_pb.ts
// @generated by protoc-gen-es v2.16.0 with parameter "target=ts,import_extension=none"
// @generated from file zitadel/settings/v2/legal_settings.proto (package zitadel.settings.v2, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { ResourceOwnerType } from "./settings_pb";
import { file_zitadel_settings_v2_settings } from "./settings_pb";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file zitadel/settings/v2/legal_settings.proto.
 */
export const file_zitadel_settings_v2_legal_settings: GenFile = /*@__PURE__*/
  fileDesc("Cih6aXRhZGVsL3NldHRpbmdzL3YyL2xlZ2FsX3NldHRpbmdzLnByb3RvEhN6aXRhZGVsLnNldHRpbmdzLnYyIvkBChdMZWdhbEFuZFN1cHBvcnRTZXR0aW5ncxIQCgh0b3NfbGluaxgBIAEoCRIbChNwcml2YWN5X3BvbGljeV9saW5rGAIgASgJEhEKCWhlbHBfbGluaxgDIAEoCRIVCg1zdXBwb3J0X2VtYWlsGAQgASgJEkMKE3Jlc291cmNlX293bmVyX3R5cGUYBSABKA4yJi56aXRhZGVsLnNldHRpbmdzLnYyLlJlc291cmNlT3duZXJUeXBlEhEKCWRvY3NfbGluaxgGIAEoCRITCgtjdXN0b21fbGluaxgHIAEoCRIYChBjdXN0b21fbGlua190ZXh0GAggASgJYgZwcm90bzM", [file_zitadel_settings_v2_settings]);

/**
 * @generated from message zitadel.settings.v2.LegalAndSupportSettings
 */
export type LegalAndSupportSettings = Message<"zitadel.settings.v2.LegalAndSupportSettings"> & {
  /**
   * @generated from field: string tos_link = 1;
   */
  tosLink: string;

  /**
   * @generated from field: string privacy_policy_link = 2;
   */
  privacyPolicyLink: string;

  /**
   * @generated from field: string help_link = 3;
   */
  helpLink: string;

  /**
   * @generated from field: string support_email = 4;
   */
  supportEmail: string;

  /**
   * @generated from field: zitadel.settings.v2.ResourceOwnerType resource_owner_type = 5;
   */
  resourceOwnerType: ResourceOwnerType;

  /**
   * @generated from field: string docs_link = 6;
   */
  docsLink: string;

  /**
   * @generated from field: string custom_link = 7;
   */
  customLink: string;

  /**
   * @generated from field: string custom_link_text = 8;
   */
  customLinkText: string;
};

/**
 * Describes the message zitadel.settings.v2.LegalAndSupportSettings.
 * Use `create(LegalAndSupportSettingsSchema)` to create a new message.
 */
export const LegalAndSupportSettingsSchema: GenMessage<LegalAndSupportSettings> = /*@__PURE__*/
  messageDesc(file_zitadel_settings_v2_legal_settings, 0);